Rewrote PayrollCalculator.calculate
Added functionality for holidays. If the pay interval is daily and payday lands on a weekend, it won't add a day to the output array until the day is a valid payday. If the pay interval isn't daily, people can't get paid late so if it lands on a weekend or holiday it finds the last previous valid payday. Then it resumes the mormal schedule (ie, if payday is a Monday and it's a holiday, for one interval the payday is the previous Friday then the next interval it returns to Monday.)